requirements.txt
flywheel-sdk~=15.1.0
psutil~=5.6.3
flywheel-gear-toolkit~=0.1.3
flywheel-bids~=0.9.1
